Keep Anthropic request tests in the existing provider test file
